refactor main.py into separate files encryptor.py, face_detector.py, classifier.py, video_capturer.py
